polish components & introduce types
