--- packages/components/src/radio-group/radio-context/index.js
@@ -1,11 +1,0 @@
-/**
- * WordPress dependencies
- */
-import { createContext } from '@wordpress/element';
-
-const RadioContext = createContext( {
-	state: null,
-	setState: () => {},
-} );
-
-export default RadioContext;
