+++ packages/components/src/radio-group/stories/index.story.tsx
@@ -1,32 +1,41 @@
 /**
- * WordPress dependencies
+ * External dependencies
  */
-import { useState } from '@wordpress/element';
+import type { Meta, StoryFn } from '@storybook/react';
 
 /**
  * Internal dependencies
  */
-import { Radio, RadioGroup } from '../../radio-group-ariakit/';
+import { RadioGroup } from '..';
+import { Radio } from '../radio';
+import type { RadioProps } from '../types';
 
-export default {
-	title: 'Components (Deprecated)/RadioGroup',
+/**
+ * WordPress dependencies
+ */
+import { useState } from '@wordpress/element';
+
+const meta: Meta< typeof RadioGroup > = {
 	component: RadioGroup,
-	subcomponents: { Radio },
+	title: 'Components (Deprecated)/RadioGroup (Ariakit)',
+	argTypes: {
+		// 	value: { control: { type: 'number', min: 0, max: 100, step: 1 } },
+		onChange: { action: 'onChange' },
+	},
 	parameters: {
-		docs: {
-			description: {
-				component:
-					'This component is deprecated. Use `RadioControl` or `ToggleGroupControl` instead.',
-			},
+		controls: {
+			expanded: true,
 		},
+		docs: { canvas: { sourceState: 'shown' } },
 	},
 };
+export default meta;
 
-export const _default = () => {
-	/* eslint-disable no-restricted-syntax */
+export const Default: StoryFn< typeof RadioGroup > = () => {
 	return (
 		<RadioGroup
 			// id is required for server side rendering
+			// eslint-disable-next-line no-restricted-syntax
 			id="default-radiogroup"
 			label="options"
 			defaultChecked="option2"
@@ -36,7 +45,6 @@ export const _default = () => {
 			<Radio value="option3">Option 3</Radio>
 		</RadioGroup>
 	);
-	/* eslint-enable no-restricted-syntax */
 };
 
 export const Disabled = () => {
@@ -58,7 +66,9 @@ export const Disabled = () => {
 };
 
 const ControlledRadioGroupWithState = () => {
-	const [ checked, setChecked ] = useState( 1 );
+	const [ checked, setChecked ] = useState<
+		RadioProps[ 'value' ] | undefined
+	>( 1 );
 
 	/* eslint-disable no-restricted-syntax */
 	return (
